♻️Refactor: ProjectDetail의 View 수정
프로젝트의 추가, 수정을 ProjectDetailViewController가 담당하도록 수정
